Extract value
